# seedbox/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager API",
        "description": "Simple api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "api server"
        }
    ],
    "paths": {
        "/servers": {
            "get": {
                "tags": [
                    "Servers"
                ],
                "summary": "Get list of servers",
                "description": "Returns list of servers",
                "operationId": "getServersList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servers"
                ],
                "summary": "Store server",
                "description": "Create new server",
                "operationId": "postServer",
                "requestBody": {
                    "description": "server object",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Server"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/servers/{server}": {
            "get": {
                "tags": [
                    "Servers"
                ],
                "summary": "Get server information",
                "description": "Returns server data",
                "operationId": "getServerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servers"
                ],
                "summary": "Update server",
                "description": "update an server",
                "operationId": "putServer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "server object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Server"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servers"
                ],
                "summary": "Delete server",
                "description": "delete an server",
                "operationId": "deleteServer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs user into system",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin@test.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/x-www-form-urlencoded": {},
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out current logged in user session",
                "operationId": "logout",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Server": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "url": {},
                            "name": {},
                            "status": {
                                "type": "string",
                                "default": "New",
                                "enum": [
                                    "Up",
                                    "Warning",
                                    "Down",
                                    "New"
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Bearer YOUR_TOKEN",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}